refactor blogList and tagList
