+++ src/templates/blogListTemplate.tsx
@@ -1,21 +1,20 @@
 import React, { useEffect, useState } from 'react';
-import { Link } from 'gatsby';
-import { BlogDateIcon, BlogTimeIcon } from '../components/blogIcon';
-import parse from 'html-react-parser';
-import * as JsSearch from 'js-search';
+import { Search } from 'js-search';
+import { Node } from 'types/markdown';
+import TagList from 'components/tagList';
+import BlogList from 'components/blogList';
 import Seo from '../components/seo';
-import { FrontMatter, Node } from 'types/markdown';
 
 interface BlogListPageProps {
   pageContext: {
-    nodes: Node[]
-  }
+    nodes: Node[];
+  };
 }
 
 const BlogListPage = ({ pageContext }: BlogListPageProps) => {
-  const { nodes } = pageContext
+  const { nodes } = pageContext;
 
-  const [search, setSearch] = useState(null);
+  const [search, setSearch] = useState<Search>();
   const [term, setTerm] = useState('');
   const [queryResult, setQueryResult] = useState<Node[]>([]);
   const [selectedTag, setSelectedTag] = useState('');
@@ -23,7 +22,9 @@ const BlogListPage = ({ pageContext }: BlogListPageProps) => {
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const searchTerm = e.target.value;
     let result: Node[] =
-      searchTerm === '' ? nodes : search.search(searchTerm);
+      searchTerm === '' || !search
+        ? nodes
+        : (search.search(searchTerm) as Node[]);
 
     if (selectedTag) {
       result = result.filter((node: Node) =>
@@ -36,7 +37,7 @@ const BlogListPage = ({ pageContext }: BlogListPageProps) => {
   };
 
   const handleClick = (e: React.MouseEvent<HTMLButtonElement>) => {
-    const buttonValue = (e.target as HTMLButtonElement).value
+    const buttonValue = (e.target as HTMLButtonElement).value;
     const clickedTag = buttonValue === selectedTag ? '' : buttonValue;
     const filteredResult =
       clickedTag === ''
@@ -50,14 +51,14 @@ const BlogListPage = ({ pageContext }: BlogListPageProps) => {
   };
 
   useEffect(() => {
-    const blogs = nodes;
-    const jsSearch = new JsSearch.Search('id');
+    // Instantiate jsSearch
+    const jsSearch = new Search('id');
 
     jsSearch.addIndex(['frontmatter', 'title']);
-    jsSearch.addDocuments(blogs);
+    jsSearch.addDocuments(nodes);
 
     setSearch(jsSearch);
-    setQueryResult(blogs);
+    setQueryResult(nodes);
   }, [nodes]);
 
   return (
@@ -88,76 +89,6 @@ const BlogListPage = ({ pageContext }: BlogListPageProps) => {
           />
         ))}
       </section>
-    </div>
-  );
-};
-
-interface TagListProps {
-  nodes: Node[];
-  selectedTag: string;
-  onClick: (e: React.MouseEvent<HTMLButtonElement>) => void
-}
-
-const TagList = ({ nodes, selectedTag, onClick }: TagListProps) => {
-  const tagSet = new Set<string>();
-  nodes.forEach((node) =>
-    node.frontmatter.tags.forEach((tag) => tagSet.add(tag))
-  );
-  const tagList = [...tagSet].sort();
-
-  return (
-    <div className="blog-tag-filter-container">
-      {tagList &&
-        tagList.map((tag, index) => {
-          const isSelected = selectedTag === tag ? 'active' : '';
-          return (
-            <button
-              className={`blog-tag blog-tag-filter ${isSelected}`}
-              value={tag}
-              onClick={onClick}
-              key={index}
-            >
-              {tag}
-            </button>
-          );
-        })}
-    </div>
-  );
-};
-
-interface BlogListProps {
-  timeToRead: number;
-  frontmatter: FrontMatter;
-  html: string
-}
-
-const BlogList = ({ timeToRead, frontmatter, html }: BlogListProps) => {
-  const htmlText = parse(html);
-
-  return (
-    <div className="blog-list-item-wrapper">
-      <Link to={'/blog/' + frontmatter.slug}>
-        <p className="blog-title ">{frontmatter.title}</p>
-      </Link>
-      <div className="blog-date-container">
-        <BlogDateIcon date={frontmatter.date} />
-        <BlogTimeIcon timeToRead={timeToRead} />
-      </div>
-      <div className="blog-tag-container">
-        {frontmatter?.tags?.map((tag, index) => (
-          <span key={index} className="blog-tag">
-            {tag}
-          </span>
-        ))}
-      </div>
-      {htmlText instanceof Array && htmlText.length > 0 && (
-        <div className="blog-content-preview">
-          {htmlText[0]}
-          <Link to={'/blog/' + frontmatter.slug}>
-            <p>Read more...</p>
-          </Link>
-        </div>
-      )}
     </div>
   );
 };
